@@ -20,11 +20,11 @@
  * SOFTWARE.
  */
 
-package org.shadowice.flocke.andotp.View;
+package org.shadowice.flocke.andotp.Dialogs;
 
 import android.app.AlertDialog;
 import android.content.DialogInterface;
-import android.support.v7.app.AppCompatDelegate;
+import androidx.appcompat.app.AppCompatDelegate;
 import android.text.Editable;
 import android.text.TextUtils;
 import android.text.TextWatcher;
@@ -45,6 +45,8 @@ import org.shadowice.flocke.andotp.Database.Entry;
 import org.shadowice.flocke.andotp.R;
 import org.shadowice.flocke.andotp.Utilities.Settings;
 import org.shadowice.flocke.andotp.Utilities.TokenCalculator;
+import org.shadowice.flocke.andotp.View.EntriesCardAdapter;
+import org.shadowice.flocke.andotp.View.TagsAdapter;
 
 import java.util.HashMap;
 import java.util.List;
@@ -58,7 +60,11 @@ public class ManualEntryDialog {
         ViewGroup container = callingActivity.findViewById(R.id.main_content);
         View inputView = callingActivity.getLayoutInflater().inflate(R.layout.dialog_manual_entry, container, false);
 
+        if (settings.getBlockAccessibility())
+            inputView.setImportantForAccessibility(View.IMPORTANT_FOR_ACCESSIBILITY_NO_HIDE_DESCENDANTS);
+
         final Spinner typeInput = inputView.findViewById(R.id.manual_type);
+        final EditText issuerInput = inputView.findViewById(R.id.manual_issuer);
         final EditText labelInput = inputView.findViewById(R.id.manual_label);
         final EditText secretInput = inputView.findViewById(R.id.manual_secret);
         final EditText counterInput = inputView.findViewById(R.id.manual_counter);
@@ -192,28 +198,28 @@ public class ManualEntryDialog {
                         Entry.OTPType type = (Entry.OTPType) typeInput.getSelectedItem();
                         TokenCalculator.HashAlgorithm algorithm = (TokenCalculator.HashAlgorithm) algorithmInput.getSelectedItem();
 
+                        String issuer = issuerInput.getText().toString();
                         String label = labelInput.getText().toString();
-                        String secret = secretInput.getText().toString();
+                        //Replace spaces with empty characters
+                        String secret = secretInput.getText().toString().replaceAll("\\s+","");
                         int digits = Integer.parseInt(digitsInput.getText().toString());
 
                         if (type == Entry.OTPType.TOTP || type == Entry.OTPType.STEAM) {
                             int period = Integer.parseInt(periodInput.getText().toString());
 
-                            Entry e = new Entry(type, secret, period, digits, label, algorithm, tagsAdapter.getActiveTags());
+                            Entry e = new Entry(type, secret, period, digits, issuer, label, algorithm, tagsAdapter.getActiveTags());
                             e.updateOTP();
                             e.setLastUsed(System.currentTimeMillis());
                             adapter.addEntry(e);
-                            adapter.saveEntries();
 
                             callingActivity.refreshTags();
                         } else if (type == Entry.OTPType.HOTP) {
                             long counter = Long.parseLong(counterInput.getText().toString());
 
-                            Entry e = new Entry(type, secret, counter, digits, label, algorithm, tagsAdapter.getActiveTags());
+                            Entry e = new Entry(type, secret, counter, digits, issuer, label, algorithm, tagsAdapter.getActiveTags());
                             e.updateOTP();
                             e.setLastUsed(System.currentTimeMillis());
                             adapter.addEntry(e);
-                            adapter.saveEntries();
                         }
                     }
                 })
